requirements.txt
torch>=2.0
torchvision>=0.15
pillow>=10.0
numpy>=1.24
matplotlib>=3.7
flask>=3.0
